% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finder.R
\name{fjord_finder}
\alias{fjord_finder}
\title{Fjord Finder}
\usage{
fjord_finder(name)
}
\arguments{
\item{name}{A string is a fjord name}
}
\value{
a string sentence
}
\description{
Check whether a name belongs to a fjord in Norway or Canada
}
\author{
Heather Zurel
}
