% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fjoRds-package.R
\docType{package}
\name{fjoRds-package}
\alias{fjoRds}
\alias{fjoRds-package}
\title{fjoRds: This package finds fjords and helps people learn R}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Zeather709/fjoRds}
  \item Report bugs at \url{https://github.com/Zeather709/fjoRds/issues}
}

}
\author{
\strong{Maintainer}: Heather Zurel \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
