% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finder.R
\name{fjord_finder}
\alias{fjord_finder}
\title{Find a Fjord}
\usage{
fjord_finder(name)
}
\arguments{
\item{name}{A fjord name provided as a character string.}
}
\description{
Check whether a name belongs to a fjord in Norway or Canada.
}
\examples{
fjord_finder("Hardangerfjord")
}
