% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fjoRds-package.R
\docType{package}
\name{fjoRds-package}
\alias{fjoRds}
\alias{fjoRds-package}
\title{fjoRds: A Package To Show How To Create Packages}
\description{
A package created for the Bergen R Ladies workshop that helps us find fjords.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/popovs/fjoRds}
  \item Report bugs at \url{https://github.com/popovs/fjoRds/issues}
}

}
\author{
\strong{Maintainer}: Sarah Popov \email{sarah.popov@mail.mcgill.ca}

}
\keyword{internal}
